backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
#aiohttp==3.8.6
gitpython==3.1.40
fpdf==1.7.2
pydantic==2.5.0
python-multipart==0.0.6
pytest==7.4.3
dotenv